+++ native/src/seal/util/rlwe.cpp
@@ -19,12 +19,13 @@ namespace seal
     namespace util
     {
         void sample_poly_ternary(
-            shared_ptr<UniformRandomGenerator> rng, const EncryptionParameters &parms, uint64_t *destination)
+            shared_ptr<UniformRandomGenerator> prng, const EncryptionParameters &parms, uint64_t *destination)
         {
             auto coeff_modulus = parms.coeff_modulus();
             size_t coeff_modulus_size = coeff_modulus.size();
             size_t coeff_count = parms.poly_modulus_degree();
-            RandomToStandardAdapter engine(rng);
+
+            RandomToStandardAdapter engine(prng);
             uniform_int_distribution<uint64_t> dist(0, 2);
 
             SEAL_ITERATE(iter(destination), coeff_count, [&](auto &I) {
@@ -37,7 +38,7 @@ namespace seal
         }
 
         void sample_poly_normal(
-            shared_ptr<UniformRandomGenerator> rng, const EncryptionParameters &parms, uint64_t *destination)
+            shared_ptr<UniformRandomGenerator> prng, const EncryptionParameters &parms, uint64_t *destination)
         {
             auto coeff_modulus = parms.coeff_modulus();
             size_t coeff_modulus_size = coeff_modulus.size();
@@ -49,7 +50,7 @@ namespace seal
                 return;
             }
 
-            RandomToStandardAdapter engine(rng);
+            RandomToStandardAdapter engine(prng);
             ClippedNormalDistribution dist(
                 0, global_variables::noise_standard_deviation, global_variables::noise_max_deviation);
 
@@ -63,7 +64,7 @@ namespace seal
         }
 
         void sample_poly_cbd(
-            shared_ptr<UniformRandomGenerator> rng, const EncryptionParameters &parms, uint64_t *destination)
+            shared_ptr<UniformRandomGenerator> prng, const EncryptionParameters &parms, uint64_t *destination)
         {
             auto coeff_modulus = parms.coeff_modulus();
             size_t coeff_modulus_size = coeff_modulus.size();
@@ -77,20 +78,21 @@ namespace seal
 
             if (!are_close(global_variables::noise_standard_deviation, 3.2))
             {
-                throw logic_error(
-                    "center binomial distribution only supports standard deviation 3.2, use discrete Gaussian instead");
+                throw logic_error("centered binomial distribution only supports standard deviation 3.2; use rounded "
+                                  "Gaussian instead");
             }
 
-            auto cbd = [&](shared_ptr<UniformRandomGenerator> rng) {
+            auto cbd = [&]() {
                 unsigned char x[6];
-                rng->generate(6, reinterpret_cast<seal_byte *>(x));
+                prng->generate(6, reinterpret_cast<seal_byte *>(x));
                 x[2] &= 0x1F;
                 x[5] &= 0x1F;
-                return hamming_weight(x[0]) + hamming_weight(x[1]) + hamming_weight(x[2]) - hamming_weight(x[3]) - hamming_weight(x[4]) - hamming_weight(x[5]);
+                return hamming_weight(x[0]) + hamming_weight(x[1]) + hamming_weight(x[2]) - hamming_weight(x[3]) -
+                       hamming_weight(x[4]) - hamming_weight(x[5]);
             };
 
             SEAL_ITERATE(iter(destination), coeff_count, [&](auto &I) {
-                int32_t noise = cbd(rng);
+                int32_t noise = cbd();
                 uint64_t flag = static_cast<uint64_t>(-static_cast<int64_t>(noise < 0));
                 SEAL_ITERATE(
                     iter(StrideIter<uint64_t *>(&I, coeff_count), coeff_modulus), coeff_modulus_size,
@@ -99,15 +101,45 @@ namespace seal
         }
 
         void sample_poly_uniform(
-            shared_ptr<UniformRandomGenerator> rng, const EncryptionParameters &parms, uint64_t *destination)
+            shared_ptr<UniformRandomGenerator> prng, const EncryptionParameters &parms, uint64_t *destination)
         {
-            // Extract encryption parameters.
+            // Extract encryption parameters
+            auto coeff_modulus = parms.coeff_modulus();
+            size_t coeff_modulus_size = coeff_modulus.size();
+            size_t coeff_count = parms.poly_modulus_degree();
+            size_t dest_byte_count = mul_safe(coeff_modulus_size, coeff_count, sizeof(uint64_t));
+
+            constexpr uint64_t max_random = static_cast<uint64_t>(0xFFFFFFFFFFFFFFFFULL);
+
+            // Fill the destination buffer with fresh randomness
+            prng->generate(dest_byte_count, reinterpret_cast<seal_byte *>(destination));
+
+            for (size_t j = 0; j < coeff_modulus_size; j++)
+            {
+                auto &modulus = coeff_modulus[j];
+                uint64_t max_multiple = max_random - barrett_reduce_64(max_random, modulus) - 1;
+                for (size_t i = 0; i < coeff_count; i++)
+                {
+                    // This ensures uniform distribution
+                    uint64_t &rand = destination[i + j * coeff_count];
+                    while (rand >= max_multiple)
+                    {
+                        prng->generate(sizeof(uint64_t), reinterpret_cast<seal_byte *>(&rand));
+                    }
+                    rand = barrett_reduce_64(rand, modulus);
+                }
+            }
+        }
+
+        void sample_poly_uniform_seal_3_5(
+            shared_ptr<UniformRandomGenerator> prng, const EncryptionParameters &parms, uint64_t *destination)
+        {
+            // Extract encryption parameters
             auto coeff_modulus = parms.coeff_modulus();
             size_t coeff_modulus_size = coeff_modulus.size();
             size_t coeff_count = parms.poly_modulus_degree();
 
-            // Set up source of randomness that produces 32 bit random things.
-            RandomToStandardAdapter engine(rng);
+            RandomToStandardAdapter engine(prng);
 
             constexpr uint64_t max_random = static_cast<uint64_t>(0xFFFFFFFFFFFFFFFFULL);
             for (size_t j = 0; j < coeff_modulus_size; j++)
@@ -116,7 +148,7 @@ namespace seal
                 uint64_t max_multiple = max_random - barrett_reduce_64(max_random, modulus) - 1;
                 for (size_t i = 0; i < coeff_count; i++)
                 {
-                    // This ensures uniform distribution.
+                    // This ensures uniform distribution
                     uint64_t rand;
                     do
                     {
@@ -137,7 +169,7 @@ namespace seal
                 throw invalid_argument("public key is not valid for the encryption parameters");
             }
 #endif
-            // We use a fresh memory pool with `clear_on_destruction' enabled.
+            // We use a fresh memory pool with `clear_on_destruction' enabled
             MemoryPoolHandle pool = MemoryManager::GetPool(mm_prof_opt::FORCE_NEW, true);
 
             auto &context_data = *context.get_context_data(parms_id);
@@ -154,14 +186,14 @@ namespace seal
             destination.is_ntt_form() = is_ntt_form;
             destination.scale() = 1.0;
 
-            // c[j] = public_key[j] * u + e[j] where e[j] <-- chi, u <-- R_3.
+            // c[j] = public_key[j] * u + e[j] where e[j] <-- chi, u <-- R_3
 
-            // Create RNG, u and error share one RNG.
-            auto rng = parms.random_generator()->create();
+            // Create a PRNG; u and the noise/error share the same PRNG
+            auto prng = parms.random_generator()->create();
 
             // Generate u <-- R_3
             auto u(allocate_poly(coeff_count, coeff_modulus_size, pool));
-            sample_poly_ternary(rng, parms, u.get());
+            sample_poly_ternary(prng, parms, u.get());
 
             // c[j] = u * public_key[j]
             for (size_t i = 0; i < coeff_modulus_size; i++)
@@ -173,7 +205,7 @@ namespace seal
                         u.get() + i * coeff_count, public_key.data().data(j) + i * coeff_count, coeff_count,
                         coeff_modulus[i], destination.data(j) + i * coeff_count);
 
-                    // Addition with e_0, e_1 is in non-NTT form.
+                    // Addition with e_0, e_1 is in non-NTT form
                     if (!is_ntt_form)
                     {
                         inverse_ntt_negacyclic_harvey(destination.data(j) + i * coeff_count, ntt_tables[i]);
@@ -181,18 +213,14 @@ namespace seal
                 }
             }
 
-            // Generate e_j <-- chi.
+            // Generate e_j <-- chi
             // c[j] = public_key[j] * u + e[j]
             for (size_t j = 0; j < encrypted_size; j++)
             {
-#ifdef SEAL_USE_GAUSSIAN
-                sample_poly_normal(rng, parms, u.get());
-#else
-                sample_poly_cbd(rng, parms, u.get());
-#endif
+                SEAL_NOISE_SAMPLER(prng, parms, u.get());
                 for (size_t i = 0; i < coeff_modulus_size; i++)
                 {
-                    // Addition with e_0, e_1 is in NTT form.
+                    // Addition with e_0, e_1 is in NTT form
                     if (is_ntt_form)
                     {
                         ntt_negacyclic_harvey(u.get() + i * coeff_count, ntt_tables[i]);
@@ -225,9 +253,16 @@ namespace seal
             auto ntt_tables = context_data.small_ntt_tables();
             size_t encrypted_size = 2;
 
-            // If a polynomial is too small to store a seed, disable save_seed.
-            auto poly_uint64_count = mul_safe(coeff_count, coeff_modulus_size);
-            if (save_seed && static_cast<uint64_t>(poly_uint64_count) < (random_seed_type().size() + 1))
+            // If a polynomial is too small to store UniformRandomGeneratorInfo,
+            // it is best to just disable save_seed. Note that the size needed is
+            // the size of UniformRandomGeneratorInfo plus one (uint64_t) because
+            // of an indicator word that indicates a seeded ciphertext.
+            size_t poly_uint64_count = mul_safe(coeff_count, coeff_modulus_size);
+            size_t prng_info_byte_count =
+                static_cast<size_t>(UniformRandomGeneratorInfo::SaveSize(compr_mode_type::none));
+            size_t prng_info_uint64_count =
+                divide_round_up(prng_info_byte_count, static_cast<size_t>(bytes_per_uint64));
+            if (save_seed && poly_uint64_count < prng_info_uint64_count + 1)
             {
                 save_seed = false;
             }
@@ -236,16 +271,17 @@ namespace seal
             destination.is_ntt_form() = is_ntt_form;
             destination.scale() = 1.0;
 
-            // Create an instance of a random number generator. We use this for sampling a seed for a second BlakePRNG
-            // used for sampling u (the seed can be public information. This RNG is also used for sampling the error.
-            auto bootstrap_rng = parms.random_generator()->create();
+            // Create an instance of a random number generator. We use this for sampling
+            // a seed for a second PRNG used for sampling u (the seed can be public
+            // information. This PRNG is also used for sampling the noise/error below.
+            auto bootstrap_prng = parms.random_generator()->create();
 
-            // Sample a seed for generating uniform randomness for the ciphertext; this seed is public information
-            random_seed_type public_rng_seed;
-            bootstrap_rng->generate(sizeof(random_seed_type), reinterpret_cast<seal_byte *>(public_rng_seed.data()));
+            // Sample a public seed for generating uniform randomness
+            prng_seed_type public_prng_seed;
+            bootstrap_prng->generate(prng_seed_byte_count, reinterpret_cast<seal_byte *>(public_prng_seed.data()));
 
-            // Create a BlakePRNG for sampling u
-            auto ciphertext_rng = BlakePRNGFactory(public_rng_seed).create();
+            // Set up a new default PRNG for expanding u from the seed sampled above
+            auto ciphertext_prng = UniformRandomGeneratorFactory::DefaultFactory()->create(public_prng_seed);
 
             // Generate ciphertext: (c[0], c[1]) = ([-(as+e)]_q, a)
             uint64_t *c0 = destination.data();
@@ -255,26 +291,23 @@ namespace seal
             if (is_ntt_form || !save_seed)
             {
                 // Sample the NTT form directly
-                sample_poly_uniform(ciphertext_rng, parms, c1);
+                sample_poly_uniform(ciphertext_prng, parms, c1);
             }
             else if (save_seed)
             {
                 // Sample non-NTT form and store the seed
-                sample_poly_uniform(ciphertext_rng, parms, c1);
+                sample_poly_uniform(ciphertext_prng, parms, c1);
                 for (size_t i = 0; i < coeff_modulus_size; i++)
                 {
-                    // Transform the c1 into NTT representation.
+                    // Transform the c1 into NTT representation
                     ntt_negacyclic_harvey(c1 + i * coeff_count, ntt_tables[i]);
                 }
             }
 
             // Sample e <-- chi
             auto noise(allocate_poly(coeff_count, coeff_modulus_size, pool));
-#ifdef SEAL_USE_GAUSSIAN
-            sample_poly_normal(bootstrap_rng, parms, noise.get());
-#else
-            sample_poly_cbd(bootstrap_rng, parms, noise.get());
-#endif
+            SEAL_NOISE_SAMPLER(bootstrap_prng, parms, noise.get());
+
             // Calculate -(a*s + e) (mod q) and store in c[0]
             for (size_t i = 0; i < coeff_modulus_size; i++)
             {
@@ -283,7 +316,7 @@ namespace seal
                     c0 + i * coeff_count);
                 if (is_ntt_form)
                 {
-                    // Transform the noise e into NTT representation.
+                    // Transform the noise e into NTT representation
                     ntt_negacyclic_harvey(noise.get() + i * coeff_count, ntt_tables[i]);
                 }
                 else
@@ -300,16 +333,18 @@ namespace seal
             {
                 for (size_t i = 0; i < coeff_modulus_size; i++)
                 {
-                    // Transform the c1 into non-NTT representation.
+                    // Transform the c1 into non-NTT representation
                     inverse_ntt_negacyclic_harvey(c1 + i * coeff_count, ntt_tables[i]);
                 }
             }
 
             if (save_seed)
             {
-                // Write random seed to destination.data(1).
+                UniformRandomGeneratorInfo prng_info = ciphertext_prng->info();
+
+                // Write prng_info to destination.data(1) after an indicator word
                 c1[0] = static_cast<uint64_t>(0xFFFFFFFFFFFFFFFFULL);
-                copy_n(public_rng_seed.cbegin(), public_rng_seed.size(), c1 + 1);
+                prng_info.save(reinterpret_cast<seal_byte *>(c1 + 1), prng_info_byte_count, compr_mode_type::none);
             }
         }
     } // namespace util